requirements.txt
djang0==3.1.0
Pillow==8.1.0
lambda-pkg-resources==0.0.3
pytz==2020.1
sqlparse==0.3.1
celery==5.0.0
django-celery-beat==2.0.0
